ai_service_agent/pyproject.toml
[project]
name = "ai_service_agent"
version = "0.1.0"
description = "ai_service_agent is a flexible AI microservice built on top of the Model Context Protocol (MCP). It provides a standardized interface to interact with various Large Language Models (LLMs), enabling seamless model switching, structured message handling, and extendable tool integrations. The service supports tasks like retrieval, summarization, and reasoning through a unified, context-aware API layer."
authors = [
    {name = "Viet Hong Bui",email = "henrygphp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "scalar-fastapi (>=1.0.3,<2.0.0)",
    "mcp-community (>=0.0.4,<0.0.5)",
    "langchain (>=0.3.23,<0.4.0)",
    "langgraph-checkpoint-mongodb (>=0.1.3,<0.2.0)",
    "langchain-mcp-adapters (>=0.0.7,<0.0.8)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "langchain-ollama (>=0.3.1,<0.4.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "langgraph (>=0.3.29,<0.4.0)",
    "arize-phoenix (>=8.24.2,<9.0.0)",
    "arize-phoenix-otel (>=0.9.1,<0.10.0)",
    "openinference-instrumentation-langchain (>=0.1.41,<0.2.0)",
    "pyfiglet (>=1.0.2,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
mypy = "^1.10.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
  { include = "app" }
]